refactor(platform): auto-discover DeviceIPCWrapper subclasses under platform/<device>/
