+++ lmcache/v1/multiprocess/custom_types.py
@@ -2,15 +2,15 @@
 # Standard
 from dataclasses import dataclass, field
 from typing import Any, Callable
-import pickle
-import threading
 
 # Third Party
 import msgspec
 import torch
 
 # First Party
-from lmcache import torch_dev, torch_device_type
+from lmcache.v1.platform.base_ipc_wrapper import (  # noqa: E402,F401
+    DeviceIPCWrapper,
+)
 
 """
 Defines the types and the customized encoder/decoders for inter-process
@@ -21,230 +21,6 @@ Key Types:
   - Contains token_ids, start, end, request_id (all required)
   - Converted to ObjectKey for storage operations via ipc_key_to_object_keys()
 """
-
-
-class DeviceIPCWrapper:
-    """Base class for KV-cache IPC wrapper.
-
-    Holds the device-agnostic mechanism shared by all transports: the
-    interface fields (``dtype``/``shape``/``stride``/``storage_offset``/
-    ``device_uuid``), UUID<->ordinal discovery via the ``torch_dev``
-    abstraction, equality, and pickle-based (de)serialization.
-
-    Every wire-level wrapper subclasses this so they share the single
-    msgspec ext code (1) registered for ``DeviceIPCWrapper``: pickle
-    preserves the concrete subclass identity across the wire so
-    ``to_tensor`` dispatches correctly on the receiving side.
-
-    Subclasses implement ``__init__`` (populate the interface fields from a
-    tensor) and ``to_tensor`` (reconstruct the tensor from the handle).
-    """
-
-    _discovered_device_mapping: dict[str, int] = {}
-    _device_mapping_lock = threading.Lock()
-
-    @classmethod
-    def _get_device_uuid(cls, device_index: int) -> str:
-        """Get the UUID of a device given its index."""
-        return str(torch_dev.get_device_properties(device_index).uuid)
-
-    @classmethod
-    def _discover_devices(cls):
-        """Discover all available accelerator devices and map their UUIDs
-        to the physical device ordinals.
-        """
-        if not torch_dev.is_available():
-            return
-
-        num_devices = torch_dev.device_count()
-        with DeviceIPCWrapper._device_mapping_lock:
-            if DeviceIPCWrapper._discovered_device_mapping:
-                return  # Already discovered
-
-            for i in range(num_devices):
-                device_uuid = cls._get_device_uuid(i)
-                DeviceIPCWrapper._discovered_device_mapping[device_uuid] = i
-
-    @classmethod
-    def _get_device_index_from_uuid(cls, device_uuid: str) -> int:
-        """Get the physical device ordinal from its UUID."""
-        cls._discover_devices()
-
-        with DeviceIPCWrapper._device_mapping_lock:
-            device_index = DeviceIPCWrapper._discovered_device_mapping.get(
-                device_uuid, None
-            )
-
-        if device_index is None:
-            raise RuntimeError(
-                f"Device UUID {device_uuid} not found in the discovered "
-                "devices. Please make sure the process can see all the "
-                "accelerator devices"
-            )
-        return device_index
-
-    def to_tensor(self) -> torch.Tensor:
-        """Reconstruct the tensor in this process from the IPC handle.
-
-        Subclasses implement the transport-specific reconstruction.
-        """
-        raise NotImplementedError
-
-    def __eq__(self, other):
-        if type(self) is not type(other):
-            return False
-        return (
-            self.handle == other.handle
-            and self.dtype == other.dtype
-            and self.shape == other.shape
-            and self.stride == other.stride
-            and self.storage_offset == other.storage_offset
-            and self.device_uuid == other.device_uuid
-        )
-
-    @staticmethod
-    def Serialize(obj: "DeviceIPCWrapper") -> bytes:
-        return pickle.dumps(obj)
-
-    @staticmethod
-    def Deserialize(data: bytes) -> "DeviceIPCWrapper":
-        return pickle.loads(data)
-
-
-class CudaIPCWrapper(DeviceIPCWrapper):
-    def __init__(self, tensor: torch.Tensor):
-        # First Party
-        from lmcache.v1.gpu_connector.kv_format.contiguity import (
-            attempt_permute_to_contiguous_view,
-        )
-
-        # Permute any non-contiguous view (e.g. vLLM's NHD-over-HND) so the
-        # shape/stride we encode across IPC reflects the physical layout.
-        # Offset is preserved by the wrapper's storage_offset field.
-        tensor = attempt_permute_to_contiguous_view(tensor)
-
-        storage = tensor.untyped_storage()
-        handle = storage._share_cuda_()
-
-        self.handle = handle
-        self.dtype = tensor.dtype
-        self.shape = tuple(tensor.shape)
-        self.stride = tuple(tensor.stride())
-        self.storage_offset = int(tensor.storage_offset())
-
-        device_index = tensor.device.index
-        self.device_uuid = self._get_device_uuid(device_index)
-
-    def to_tensor(self) -> torch.Tensor:
-        """
-        Note:
-            This function may break if the accelerator is not initialized.
-            We should call `torch_dev.init()` before using this function
-            (guarded by hasattr since not all backends expose init()).
-        """
-        device_index = self._get_device_index_from_uuid(self.device_uuid)
-
-        storage = torch.UntypedStorage._new_shared_cuda(  # noqa: SLF001
-            device_index, *self.handle[1:]
-        )
-
-        t = torch.empty(
-            (), device=f"{torch_device_type}:{device_index}", dtype=self.dtype
-        )
-        t.set_(storage, self.storage_offset, self.shape, self.stride)
-        return t
-
-
-class RawCudaIPCWrapper(DeviceIPCWrapper):
-    """IPC wrapper for CUDA tensors allocated outside PyTorch's caching
-    allocator.
-
-    PyTorch's ``UntypedStorage._share_cuda_()`` only works for tensors
-    backed by its own caching allocator. TRT-LLM publishes its KV pool
-    via ``at::for_blob`` over a ``cudaMalloc``'d buffer, which raises in
-    ``_share_cuda_()``. This subclass bypasses that path: it calls
-    ``cudaIpcGetMemHandle`` on the raw data pointer, then reconstructs
-    the tensor on the receiving side via ``cudaIpcOpenMemHandle`` plus
-    a CuPy ``UnownedMemory`` → DLPack → ``torch`` round-trip.
-
-    Sharing the ``DeviceIPCWrapper`` base (rather than introducing a
-    parallel class with its own msgspec ext code) is load-bearing —
-    msgspec does not support unions of custom ext-encoded types. With a
-    common base, ``KVCache = list[DeviceIPCWrapper]`` type-checks, the
-    single ext code 1 round-trips every wrapper, and pickle preserves
-    the concrete subclass identity through the wire so ``to_tensor``
-    dispatches correctly.
-    """
-
-    def __init__(self, tensor: torch.Tensor) -> None:
-        # First Party
-        from lmcache.v1.gpu_connector.utils import assert_contiguous
-
-        assert_contiguous(tensor)
-
-        try:
-            # Third Party
-            from cuda.bindings import runtime as cudart
-        except ImportError:
-            # Third Party
-            from cuda import cudart
-
-        data_ptr = tensor.data_ptr()
-        err, ipc_handle = cudart.cudaIpcGetMemHandle(data_ptr)
-        if err != cudart.cudaError_t.cudaSuccess:
-            raise RuntimeError(
-                f"cudaIpcGetMemHandle failed: {err} (ptr=0x{data_ptr:x})"
-            )
-
-        # Store only what's needed for reconstruction.
-        self._ipc_handle_reserved = bytes(ipc_handle.reserved)
-        self._nbytes = tensor.untyped_storage().nbytes()
-
-        # DeviceIPCWrapper interface fields. ``handle`` is unused —
-        # ``to_tensor`` is overridden to bypass it — but kept (None) so
-        # the base-class equality check has a value to compare.
-        self.handle = None
-        self.dtype = tensor.dtype
-        self.shape = tuple(tensor.shape)
-        self.stride = tuple(tensor.stride())
-        self.storage_offset = int(tensor.storage_offset())
-
-        device_index = tensor.device.index
-        self.device_uuid = self._get_device_uuid(device_index)
-
-    def to_tensor(self) -> torch.Tensor:
-        """Reconstruct the tensor in this process via raw CUDA IPC."""
-        # Third Party
-        import cupy
-
-        try:
-            # Third Party
-            from cuda.bindings import runtime as cudart
-        except ImportError:
-            # Third Party
-            from cuda import cudart
-
-        device_index = self._get_device_index_from_uuid(self.device_uuid)
-
-        handle = cudart.cudaIpcMemHandle_t()
-        handle.reserved = self._ipc_handle_reserved
-        err, ptr = cudart.cudaIpcOpenMemHandle(
-            handle, cudart.cudaIpcMemLazyEnablePeerAccess
-        )
-        if err != cudart.cudaError_t.cudaSuccess:
-            raise RuntimeError(f"cudaIpcOpenMemHandle failed: {err}")
-
-        # Wrap as a flat ``uint8`` CuPy array, DLPack to torch, then view
-        # as the original dtype/shape. ``uint8`` avoids dtype-conversion
-        # gaps (bfloat16, fp8 have no direct CuPy/NumPy equivalent without
-        # ml_dtypes).
-        with cupy.cuda.Device(device_index):
-            mem = cupy.cuda.UnownedMemory(ptr, self._nbytes, owner=self)
-            memptr = cupy.cuda.MemoryPointer(mem, 0)
-            cp_flat = cupy.ndarray(self._nbytes, dtype=cupy.uint8, memptr=memptr)
-
-        raw = torch.from_dlpack(cp_flat)
-        return raw.view(self.dtype).reshape(self.shape)
 
 
 @dataclass(order=True, frozen=True)
@@ -432,6 +208,17 @@ class CBUnifiedLookupResult:
     non_prefix_segments: list[CBMatchResult]
     segmented_prefix_segments: list[CBMatchResult] = field(default_factory=list)
 
+
+# ------------------------------------------------------------------
+# Backward-compatibility re-exports: the canonical home for
+# ``CudaIPCWrapper`` / ``RawCudaIPCWrapper`` is now
+# ``lmcache.v1.platform.cuda.ipc_wrapper``.
+# ------------------------------------------------------------------
+# First Party
+from lmcache.v1.platform.cuda.ipc_wrapper import (  # noqa: E402,F401
+    CudaIPCWrapper,
+    RawCudaIPCWrapper,
+)
 
 _CUSTOMERIZED_SERIALIZERS = {
     DeviceIPCWrapper: CustomizedSerdeConfig(
