+++ lmcache/v1/platform/cuda/__init__.py
@@ -1,35 +1,17 @@
 # SPDX-License-Identifier: Apache-2.0
 """CUDA-specific platform primitives.
 
-Importing this package self-registers a CUDA KV-cache wrapper
-factory with :mod:`lmcache.v1.platform._registry`, so the dispatch
-in :mod:`lmcache.integration.vllm.vllm_multi_process_adapter` can
-locate it by ``tensor.device.type``.
+:class:`~lmcache.v1.platform.cuda.ipc_wrapper.CudaIPCWrapper` carries
+a ``device_type`` ClassVar and a ``wrap`` factory classmethod, which
+:func:`~lmcache.v1.platform._registry._discover_wrappers_once` picks
+up at run-time -- no static ``register_kv_wrapper`` needed.
+
+The CUDA availability predicate is still registered statically here
+so callers can check ``is_available("cuda")`` at import time.
 """
 
-# Standard
-from typing import Any
-
-# Third Party
-import torch
-
 # First Party
-from lmcache.v1.platform._registry import (
-    register_availability,
-    register_kv_wrapper,
-)
-
-
-def _kv_wrapper_factory(tensor: torch.Tensor) -> Any:
-    """Indirect-dispatch wrapper.
-
-    Re-imports :class:`CudaIPCWrapper` on every call so test suites
-    that swap the symbol still see their override take effect.
-    """
-    # First Party
-    from lmcache.v1.multiprocess.custom_types import CudaIPCWrapper
-
-    return CudaIPCWrapper(tensor)
+from lmcache.v1.platform._registry import register_availability
 
 
 def _cuda_is_available() -> bool:
@@ -41,4 +23,3 @@ def _cuda_is_available() -> bool:
 
 
 register_availability("cuda", _cuda_is_available)
-register_kv_wrapper("cuda", _kv_wrapper_factory)
